+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `ElevenLabsTTSService` now supports additional runtime changes to the `model`
-  and `language` parameters.
+- `ElevenLabsTTSService` now supports additional runtime changes to the `model`,
+  `language`, and `voice_settings` parameters.
 
 - Added `MistralLLMService`, using Mistral's chat completion API.